chore(ContextMenuSelectionInfo): Move to own file
Refs: #12149
